+++ docs/source/changelog.rst
@@ -1,4 +1,4 @@
-.. Created by changelog.py at 2022-04-13, command
+.. Created by changelog.py at 2022-05-06, command
    '/Users/giffler/.cache/pre-commit/repor6pnmwlm/py_env-default/bin/changelog docs/source/changes compile --output=docs/source/changelog.rst'
    based on the format of 'https://keepachangelog.com/'
 
@@ -6,13 +6,14 @@
 CHANGELOG
 #########
 
-[Unreleased] - 2022-04-13
+[Unreleased] - 2022-05-06
 =========================
 
 Added
 -----
 
 * Introduce a TARDIS REST API to query the state of resources from SqlRegistry
+* Add support for passing environment variables as executable arguments to support HTCondor grid universe
 
 Changed
 -------